Check if device is keyboard (instead of not mouse)
The program seems to recognize several devices as mice even when they
are keyboards. Maybe they are both for some reason?
In any case, for me this is fixed by checking if the device
conforms to a keyboard, instead of excluding devices that conform
to mice. It can't be that simple though, right...? But it does not
seem to pick up events from my Apple mouse or trackpad.